Alex.Reto6 Un programa que contiene un pequeño videojuego de adivinar la palabra contiene 3 niveles de dificultad además de agregar dos excepciónes a pósibles errores por parte del usuario
