+++ JAVA/src/edu/alex/reto6/Main.java
@@ -1,6 +1,6 @@
-package edu.alex.reto5;
+package edu.alex.reto6;
 
-import edu.alex.reto5.ui.CLI;
+import edu.alex.reto6.ui.CLI;
 
 /**
  * Es la clase main donde simplemente se manda a llmar la clase CLI para
